♻️ Move notification permission code to VM
