+++ HRHN/ViewModel/TodayViewModel.swift
@@ -5,13 +5,14 @@
 //  Created by Chanhee Jeong on 2022/12/25.
 //
 
-import Foundation
+import UIKit
 
 final class TodayViewModel {
     
     var todayChallenge: Observable<String?> = Observable(nil)
     
     private var coreDataManager = CoreDataManager.shared
+    private let center = UNUserNotificationCenter.current()
     
     init(){}
     
@@ -21,6 +22,15 @@ final class TodayViewModel {
             self.todayChallenge = Observable(challenges[0].content)
         } else {
             self.todayChallenge = Observable(nil)
+        }
+    }
+    
+    func requestNotificationAuthorization() {
+        let authOptions = UNAuthorizationOptions(arrayLiteral: .alert, .badge, .sound)
+        center.requestAuthorization(options: authOptions) { success, error in
+            if let error = error {
+                print("Auth Error: ", error)
+            }
         }
     }
 }
